pyproject.toml
[project]
name = "lasergrid-render"
version = "0.1.0"
description = "Automation helpers to procedurally render Blender dioramas"
authors = [{ name = "LaserGrid" }]
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
  "pathvalidate>=3.2",
  "fake-bpy-module-4-2"
]

[build-system]
requires = ["setuptools>=65"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
